% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.text.R
\name{add.text}
\alias{add.text}
\title{Annotate Plots.}
\usage{
add.text(cex = 1)
}
\arguments{
\item{cex}{provides control over the text font size (default=1).}
}
\value{
Draft \code{\link{text}} commands that might be added to a function.
}
\description{
Used to annotate the plots. Add annotation on points chosen on
}
\examples{
plot(att.stock.tsd)
add.text()

}
\seealso{
documentation for \code{\link{locator}}.
}
