% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.contour.plot.R
\name{arima.contour.plot}
\alias{arima.contour.plot}
\title{ARIMA Model Likelihood Contour Plot}
\usage{
arima.contour.plot(
  data.tsd,
  model,
  param1 = list(1, seq(-0.99, 0.99, length = 8)),
  param2 = list(2, seq(-0.99, 0.99, length = 8)),
  relative = FALSE,
  verbose = FALSE,
  my.title,
  persp.plot = FALSE,
  ...
)
}
\arguments{
\item{data.tsd}{Time series data object.}

\item{model}{Time series model.}

\item{param1}{List containing the integer parameter number of the first parameter and sequence of values at which the likelihood will be evaluated.}

\item{param2}{List containing the integer parameter number of the second parameter and sequence of values at which the likelihood will be evaluated.}

\item{relative}{If true, the relative likelihood is plotted instead of the log likelihood (default is FALSE)}

\item{verbose}{Prints model coefficient values for each evaluation point; useful to debug estimation problems when doing a profile plot (when the number of parameters is greater than 2).}

\item{my.title}{A title (or an empty character string) to replace the default title.}

\item{persp.plot}{If TRUE a perspeciive plot (instead of a contour plot) will be provided.}

\item{...}{Allows sending down extra arguments to the contour or persp functions (such as graphs parameters).}
}
\value{
Invisibly returns the information used to make the contour plot.
}
\description{
Provides a contour plot of the likelihood or two-dimensional profile likelihood for a specified ARIMA model and time series data object.
}
\examples{
#ar1=2 model
arima.contour.plot(device.inventory.tsd, model = model.pdq(p = 2),
         list(1, seq(-1.99, 1.99, length = 50)),
         list(2, seq(-.99, .99, length = 50)))
#ar1=2 model--zoom in
\dontrun{arima.contour.plot(device.inventory.tsd, model = model.pdq(p = 2),
         list(1, seq(1.35, 1.394, length = 40)),
         list(2, seq(-0.408, -0.390, length = 40)), my.title='')}

}
\seealso{
\code{\link{arima.likelihood.plot}}.
}
