% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.likelihood.plot.R
\name{arima.likelihood.plot}
\alias{arima.likelihood.plot}
\title{ARIMA Model Likelihood Plot}
\usage{
arima.likelihood.plot(data.tsd, model, param1 = list(1, seq(-0.99, 0.99,
  length = 10)), relative = FALSE, verbose = FALSE, my.title)
}
\arguments{
\item{data.tsd}{Time series data object.}

\item{model}{Time series model.}

\item{param1}{List containing the integer parameter number of the parameter and sequence of values at which the likelihood will be evaluated.}

\item{relative}{If true, the relative likelihood is plotted instead of the log likelihood (default if FALSE)}

\item{verbose}{Prints model coefficient values for each evaluation point; useful to debug estimation problems when doing a profile plot (when the number of parameters is greater than 1).}

\item{my.title}{A title (or an empty character string) to replace the default title.}
}
\value{
Invisibly returns the information used to make the contour plot.
}
\description{
Provides a plot of the likelihood or one-dimensional profile likelihood for a specified ARIMA model and times series data object.
}
\examples{
#ar1 model
arima.likelihood.plot(device.inventory.tsd, model = model.pdq(p = 1),
           list(1, seq(-0.5, .9999, length =50)))

#ar1 model zoom in
arima.likelihood.plot(device.inventory.tsd,model = model.pdq(p = 1),
           list(1, seq(0.5, .9999, length =50)), my.title='')

}
\seealso{
\code{\link{arima.contour.plot}}.
}
