% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma.roots.R
\name{arma.roots}
\alias{arma.roots}
\title{AR or MA Polynomial Roots}
\usage{
arma.roots(coeficients, nplot = 400, my.title = NULL, digits = 4)
}
\arguments{
\item{coeficients}{A vector of coefficients of polynomial.}

\item{nplot}{The number of points used to display the plot (default is 400).}

\item{my.title}{If my.title = NULL, it will display coefficient on middle top of graph; otherwise, it will not display coeeficient on graph.}

\item{digits}{The number of decimal places to be used in numeric output (default is 4).}
}
\value{
The roots of the polynomial are returned as a vector, but invisibly.
}
\description{
Compute and display an AR or an MA polynomial and its roots, relative to the 'unit circle'.
}
\examples{
arma.roots(c(1.6,0.4))
arma.roots(c(.5,-.1))
arma.roots(c(.5,-.9,.1,.5))
# trivial (first order or linear polynomial) example
arma.roots(.5)

}
