% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{att.stock.tsd}
\alias{att.stock.tsd}
\title{AT&T Stock Price}
\format{
A time series data object (created with function tsd).
}
\source{
Wall Stret Journal.  Also appeared as Case 6 in Pankrantz (1983).
}
\usage{
att.stock.tsd
}
\description{
Weekly closing price of AT&T common stock  from 1979 (in dollars) containing 52 observations.
}
\keyword{datasets}
