% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston.robberies.tsd}
\alias{boston.robberies.tsd}
\title{Boston Armed Robberies}
\format{A time series data object (created with function tsd).}
\source{
Originally analyzed by Deutch and Alt (1977).  Also appeared as Case 11 in Pankrantz (1983).
}
\usage{
boston.robberies.tsd
}
\description{
Monthly observations from January 1966 to October 1975.
}
\keyword{datasets}
