% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{change.inventory.tsd}
\alias{change.inventory.tsd}
\title{Change in Business Inventories}
\format{
A time series data object (created with function tsd).
}
\source{
Business Conditions Digest, November 1979.  Also appeared as Case 1 in Pankrantz (1983).
}
\usage{
change.inventory.tsd
}
\description{
A seasonally adjusted time series data object containing 60 observations from the first quarter of 1955 to the fourth quarter of 1969 in units of billions of dollars.
}
\keyword{datasets}
