% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.test.R
\name{color.test}
\alias{color.test}
\title{Color Test}
\usage{
color.test(n = 8)
}
\arguments{
\item{n}{is the number of colors (8 is the default because the colors recycle after col=9 in the standard R setup).}
}
\description{
Provides a pie chart showing different color options, by number.
}
\examples{
color.test()
shaded.tsplot(spot.tsd, col=4)

}
