% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum2.R
\name{cumsum2}
\alias{cumsum2}
\title{Integrate a time series}
\usage{
cumsum2(w, lag = 1, start)
}
\arguments{
\item{w}{The input time series.}

\item{lag}{Seasonal lag if a seasonal integration is desired. The default is lag=1.}

\item{start}{Optional value that will be the base value. The degault is to use the first element of w.}
}
\value{
A time series.
}
\description{
Performs the operation of integrating a time series (inverse of a difference). Allows seasonal integtation.
}
\examples{
ozone.xmat <- read.csv(RTseriesExtDataPath("ozoneRegdat.csv"))
ozo.xreg <- cbind(ozone.xmat[,2], cumsum2(ozone.xmat[,3],12), cumsum2(ozone.xmat[,4],12))
}
