% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{device.inventory.tsd}
\alias{device.inventory.tsd}
\title{Device Inventory Levels}
\format{A time series data object (created with function tsd).}
\source{
Unknown
}
\usage{
device.inventory.tsd
}
\description{
Monthly data on the inventroy levels of a device in hundreds of units. Data run from Janaury 1984 to December 1994.
}
\keyword{datasets}
