% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Durbin.R
\name{Durbin}
\alias{Durbin}
\title{Computes the pacf using the Durbin method}
\usage{
Durbin(the.acf, lag.max = 10)
}
\arguments{
\item{the.acf}{A vector giving the acf function.}

\item{lag.max}{integer saying how many pacf values to compute. The default is 10.}
}
\description{
For a given acf function, computes the pacf
}
\examples{
Durbin(c(0.4, rep(0,100)))
Durbin(0.9^(1:100))


}
