% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esti.R
\name{esti}
\alias{esti}
\title{ARIMA Model Estimation}
\usage{
esti(data.tsd, gamma = 1, m = 0, model, gof.lag = 10, lag.max = 38,
  number.forecasts = 24, pred.level = 0.95, xreg.in, y.range, x.range,
  d.trend = FALSE, print.table = FALSE, ps1 = NULL, ps2 = NULL,
  ...)
}
\arguments{
\item{data.tsd}{Input time series data object (output from the tsd function).}

\item{gamma}{Optional; the Box-Cox transformation power parameter. The default value is gamma=1 (no transformation). If a log transformation is needed, use gamma=0.}

\item{m}{Optional constant to added to the response variable before the data are transformed. The default value is m=0.}

\item{model}{A list specifying an ARIMA model. See functions model.pdq for a simple method for specifying the model input.}

\item{gof.lag}{Number of lags to be use to compute the Ljung-Box statistic (default=10).}

\item{lag.max}{The number of lags at which to estimate residual autocorrelations (default is lag.max=38).}

\item{number.forecasts}{How far into the future to forecast (default is number.forecast=24).}

\item{pred.level}{Prediction confidence level (default is pred.level =0 .95).}

\item{xreg.in}{An x matrix for dynamic regression (see \code{\link{lead.matrix}}).}

\item{y.range}{Range of the y axis for the predictions}

\item{x.range}{Range of the x axis for the predictions}

\item{d.trend}{Indicates whether a model should have a deterministic trend term or not (default is  d.trend = FALSE).}

\item{print.table}{If print.table = TRUE, a table of the forecasts and prediction intervals will be provided (default is print.table = FALSE).}

\item{ps1}{if provided, the postscript file name for the first page of graphical output}

\item{ps2}{if provided, the postscript file name for the second page of graphical output}

\item{...}{allows sending down extra arguments to the arima function (such as optimization options or fixed values of parameters).}
}
\value{
Invisibly returns the output from the R \code{\link{arima}} command used to do the estimation.
}
\description{
Performas estimation, diagnostic checking, and forecasting for a specified ARIMA model and time series data object.
}
\examples{
# AR(1) model for the sunspot data
esti(spot.tsd, model=model.pdq(p=3))

# Airline model for the international airline passengers data
Passengers.ts <- ts(Passengers, freq=12, start=1949)
Passengers.tsd <- tsd(Passengers.ts, data.title='International Airline Passengers',
                      time.units='Year', response.units='Thousands of Passengers')
airline.model <- model.pdq(period=12, d=1, D=1, q=1, Q=1)
esti(Passengers.tsd, gamma = 0, model=airline.model)
esti(Passengers.tsd, gamma = 0, m = 2, model=airline.model)

}
