% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{female.unemployment.tsd}
\alias{female.unemployment.tsd}
\title{Female Unemployment Rate}
\format{A time series data object (created with function tsd).}
\source{
Unknown
}
\usage{
female.unemployment.tsd
}
\description{
Female Unemployment Rate
}
\keyword{datasets}
