% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gas.oil.tsd}
\alias{gas.oil.tsd}
\title{Gas and Oil Consumption}
\format{A time series data object (created with function tsd).}
\source{
Unknown
}
\usage{
gas.oil.tsd
}
\description{
Quarterly values between Q1 of 1967 to Q4 of 1982.
}
\keyword{datasets}
