% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gasrx.tsd}
\alias{gasrx.tsd}
\title{Gas Furnace Input Rate}
\format{A time series data object (created with function tsd).}
\source{
Used as an example in Chapter 10 of Box and Jenkins (1970).
}
\usage{
gasrx.tsd
}
\description{
The gas inlet rate for a gas furnace.
}
\keyword{datasets}
