% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gasry.tsd}
\alias{gasry.tsd}
\title{Gas Furnace Percent CO2 Output}
\format{A time series data object (created with function tsd).}
\source{
Used in Chapter 10 of Box and Jenkins (1970)
}
\usage{
gasry.tsd
}
\description{
The percent CO2 in the output of a gas furnace.
}
\keyword{datasets}
