% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hstart}
\alias{hstart}
\title{Housing Starts}
\format{A single vector; read across each line and then down.}
\source{
U.S. Bureau of the Census, Construction Reports. Also was provided as an example in the S and S-Plus software systems.
}
\usage{
hstart
}
\description{
Monthly observations from January 1966 to December 1974.
}
\keyword{datasets}
