% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead.matrix.R
\name{lead.matrix}
\alias{lead.matrix}
\title{Leading Indicator Matrix}
\usage{
lead.matrix(xvec, lagvec, number.forecasts = 24)
}
\arguments{
\item{xvec}{A leading indicator time series}

\item{lagvec}{A vector of integers indicating the lags with non-zzero coefficients in the dynamic regression model.}

\item{number.forecasts}{The number of forecasts that will be requested (the default is 24).}
}
\value{
A time series matrix, suitable to use as input for a dynamic regression model using the esti function.
}
\description{
Uses a leading indicator time series to create a leading indicator matrix that can be used as input to the esti function.
}
\examples{
gasrx.lead.matrix <- lead.matrix(gasrx.tsd, lagvec=c(3, 4, 5, 6))
est.out <- esti(gasry.tsd, model=model.pdq(p=3), xreg.in=gasrx.lead.matrix)

}
