% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls.mode.R
\name{ls.mode}
\alias{ls.mode}
\title{List Objects of a Given Mode}
\usage{
ls.mode(pos = 1, mode = "function")
}
\arguments{
\item{pos}{Position on the search list to look (default is 1).}

\item{mode}{Type of object to report; default is function}
}
\description{
Provides a list of objects with a specified mode (functions by default).
}
\examples{
ls.mode()
ls.mode(pos=2)
ls.mode(mode='tsd')
ls.mode(mode='tsd', pos=2)

}
