% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{machine.tool.ship.tsd}
\alias{machine.tool.ship.tsd}
\title{Machine-Tool Shipments}
\format{
A time series data object (created with function tsd).
}
\source{
Survey of Current Business, U.S. Department of Commerse.  Also appeared as Case 12 in Pankrantz (1983).
}
\usage{
machine.tool.ship.tsd
}
\description{
Monthly observations from January 1968 to December 1974.
}
\keyword{datasets}
