% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{male.unemployment.tsd}
\alias{male.unemployment.tsd}
\title{Male Unemployment Rate}
\format{A time series data object (created with function tsd).}
\source{
Unknown
}
\usage{
male.unemployment.tsd
}
\description{
Male Unemployment Rate
}
\keyword{datasets}
