% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.pdq.R
\name{model.pdq}
\alias{model.pdq}
\title{Specify an ARIMA Model}
\usage{
model.pdq(p = 0, d = 0, q = 0, P = 0, D = 0, Q = 0,
  period = NULL)
}
\arguments{
\item{p}{The order of the nonseasonal AR polynomial in the model.}

\item{d}{The number of nonseasonal differences.}

\item{q}{The order of the nonseasonal MA polynomial in the model.}

\item{P}{The order of the seasonal AR polynomial in the model.}

\item{D}{The number of seasonal differences.}

\item{Q}{The order of the seasonal MA polynomial in the model.}

\item{period}{Optional parameter to allow specification of the seasonal period. If not specified, the seasonal period will be inferred from the time series data object that is specified to the \code{\link{esti}} function.}
}
\description{
Provides a simple method to specify an ARIMA model that can be used with the \code{\link{esti}} function. Defaults for the following six arguments is 0.
}
\examples{
# An AR(3) model
ar3.model <- model.pdq(p=3)
ar3.model

# The 'airline model'
airline.model <- model.pdq(period=12,d=1,D=1,q=1,Q=1)
airline.model

# The trivial model
trivial.model <- model.pdq()
trivial.model

}
