% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.acf.sim.R
\name{multiple.acf.sim}
\alias{multiple.acf.sim}
\title{Time Series Model Simulation and ACF Plot}
\usage{
multiple.acf.sim(
  model = NULL,
  sample.size.vec = c(30, 60, 120),
  number.simulations = 4
)
}
\arguments{
\item{model}{Optional list containing two components names ar and ma. These components should be vectors containing coefficianes of autoregressive and moving average parts of an ARMA model. If the list is not provided, a random model and parameters are generated.}

\item{sample.size.vec}{A vector of sample sizes. Length 3 is recommended. Default is sample.size.vec=c(30, 60, 120).}

\item{number.simulations}{Number of simulations. Default is number.simulations=4.}
}
\description{
Simulate time series realizations, with different realization sizes, from a particular time series model and display the sample ACF functions graphically.
}
\examples{
multiple.acf.sim()
multiple.acf.sim(sample.size.vec = c(50, 100, 200))

}
