% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.pi.sim.R
\name{multiple.pi.sim}
\alias{multiple.pi.sim}
\title{Prediction Intervals Simulation}
\usage{
multiple.pi.sim(sample.sizes = c(30, 120), confidence.levels = c(0.8,
  0.95), mu = 100, sigma = 10, number.simulations = 100, scale = 2)
}
\arguments{
\item{sample.sizes}{A vector of sample sizes. Length 2 is recommended. Default is sample.sizes=c(30, 120)}

\item{confidence.levels}{A vector of confidence levels. Length 2 is recommended. Default is confidence.levels=c(0.8, 0.95)}

\item{mu}{Mean of the distribution from which samples are taken (default is 100)}

\item{sigma}{Standard deviation of the distribution from which samples are taken (default is 10).}

\item{number.simulations}{Number of prediction intervals to be simulated (default is 100).}

\item{scale}{Scale factor that determines the scale of the common y-axis of the plots. The default is 2. A larger number would allow more white space around edges. A smaller number might allow some of the interval endpoints to be outside of the plots.}
}
\description{
Simulation to illustrate the effect that confidence level and sample size have on the results of prediction interval width and correctness. Simulated samples are used to compute the prediction intervals and then one additional observation is generated and plotted for each inerval. If the additional observation does not fall in the prodiction interval, the interval is marked in red.
}
\examples{
multiple.pi.sim()
multiple.pi.sim(sample.sizes = c(10, 40))
multiple.pi.sim(confidence.levels = c(0.95, 0.99))
multiple.pi.sim(sample.sizes = c(10, 40), confidence.levels = c(0.95, 0.99))

}
