% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.probplot.sim.R
\name{multiple.probplot.sim}
\alias{multiple.probplot.sim}
\title{Probability Plot Simulation}
\usage{
multiple.probplot.sim(
  distribution = "normal",
  parameter = 1,
  sample.size.vec = c(10, 20, 40),
  number.simulations = 5,
  my.title = NULL
)
}
\arguments{
\item{distribution}{Distribution from which samples are drawn.}

\item{parameter}{Value of the scale or shape parameter (depending on the distribution). Default is 1.}

\item{sample.size.vec}{Vector of sample sizes for the simulation. Defauls is c(10, 20, 40).}

\item{number.simulations}{Number of simulations for each sample size. Default is 5.}

\item{my.title}{Allows specification of an alternative title or '' for no title}
}
\description{
Simulates data and makes probability plots to allow an assessment of the expected deviation from probability plots in repeated sampling.
}
\examples{
multiple.probplot.sim()
multiple.probplot.sim(distribution='exponential')
multiple.probplot.sim(distribution='weibull', parameter=0.8)
multiple.probplot.sim(sample.size.vec=c(25, 50, 100))

}
