% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ohio.tsd}
\alias{ohio.tsd}
\title{Ohio Elecricity Consumption}
\format{A time series data object (created with function tsd).}
\source{
Used as an example in Mabert (1975).
}
\usage{
ohio.tsd
}
\description{
Ohio electricity consumption bewtween January 1954 to October 1970.
}
\keyword{datasets}
