% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ozone.tsd}
\alias{ozone.tsd}
\title{Ozone Air Polution Values}
\format{A time series data object (created with function tsd).}
\source{
Used as an example in Box and Tiao (1975)
}
\usage{
ozone.tsd
}
\description{
Ozone oxidant concertration (parts per hundred million) in downtown Los Angeles from January 1955 to December 1972.
}
\keyword{datasets}
