% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Passengers}
\alias{Passengers}
\title{Airline Passenger Numbers}
\format{
A single vector; read across each line and then down.
}
\source{
Brown (1962). Also, used as an example in Chapters 2 and 5 of Box and Jenkins (1970).
}
\usage{
Passengers
}
\description{
Monthly data on the number of international airline passengers between January 1949 and December 1960.
}
\keyword{datasets}
