% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause.R
\name{pause}
\alias{pause}
\title{Allows a pause in the running of an R function.}
\usage{
pause(
  prompt = "Touch Return to continue; type stop to breakout: ",
  skip = FALSE,
  browse = FALSE
)
}
\description{
Allows a pause in the running of an R function.
}
