% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tsd.R
\name{plot.tsd}
\alias{plot.tsd}
\title{Plot Time Series Data Object}
\usage{
\method{plot}{tsd}(x, ylab = NULL, xlab = NULL, main = NULL, ...)
}
\arguments{
\item{x}{A time series data object (created with function tsd).}

\item{ylab}{Label for the y axis. Default is response.units in data.tsd.}

\item{xlab}{Label for the time axis. Default is time.units in data.tsd.}

\item{main}{Plot title.  Default is data.title in data.tsd.}

\item{...}{Allows sending down extra arguments such as graphs parameters.}
}
\description{
Plots the time series in the specified time series data object.
}
\examples{
plot(spot.tsd)
plot(boston.robberies.tsd, las=1)
plot(gasrx.tsd)
plot(gasrx.tsd)

}
