% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.weights.R
\name{psi.weights}
\alias{psi.weights}
\title{Compute psi weights for an ARMA model}
\usage{
psi.weights(ar = numeric(), ma = numeric(), lag.max = 10)
}
\arguments{
\item{ar}{Vector of autoregressive (AR) coefficients (default is none).}

\item{ma}{Vector of moving average (MA) coefficients (default is none).}

\item{lag.max}{number of psi weights to compute}
}
\value{
A vector of lag.max coefficients.
}
\description{
Computes the coefficients of the (potentially) infinite moving average (MA) model representation of an ARMA model (using the Box-Jenkins convention for the sign of the MA coefficients). These coefficients can also be viewed as the impulse response function of the ARMA model.
}
\examples{
psi.weights(ar=0.5)
psi.weights(ar=0.5, ma=-0.5)
psi.weights(ar=0.9, ma=-0.9)
psi.weights(ar=0.99)
psi.weights(ar=0.5, ma=0.5)

}
