% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{savings.rate}
\alias{savings.rate}
\title{Personal Savings as a Percent of Disposable Personal Income}
\format{
A single vector; read across each line and then down.
}
\source{
Originally from the U.S. Department of Commerse. Also appeared as Case 2 in Pankrantz (1983).
}
\usage{
savings.rate
}
\description{
100 seasonally adjusted quarterly observations for the years 1955 to 1979.
}
\keyword{datasets}
