% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.predictions.R
\name{seasonal.predictions}
\alias{seasonal.predictions}
\title{Seasonal Prediction}
\usage{
seasonal.predictions(the.fit, the.data.frame)
}
\arguments{
\item{the.fit}{Output from the lm command fitting a dummy variable seasonal time series model.}

\item{the.data.frame}{Data frame used in itting the dummy variable seasonal time series model.}
}
\description{
Provides a plot of predictions of a linear regression model using dummy variables for seasonal effects.
}
