% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaded.tsplot.R
\name{shaded.tsplot}
\alias{shaded.tsplot}
\title{Shaded Time Series Plot}
\usage{
shaded.tsplot(x, col = 1, top = FALSE)
}
\arguments{
\item{x}{Time series object or a time series data object.}

\item{col}{The shaded color, default is 1 (black). Other options are 2 for red, 3 for green, 4 for blue, 5 for light blue, 6 for purple, 7 for yellow, and  8 for grey.}

\item{top}{If top = FALSE, the shaded part will be below the time series. If top = TRUE, the shaded part will be above the time series.}
}
\description{
Provides a shaded time series plot.
}
\examples{
plot(spot.tsd)
shaded.tsplot(spot.tsd)
shaded.tsplot(spot.tsd, col=4)
shaded.tsplot(spot.tsd, top=TRUE, col=3)

}
\seealso{
\code{\link{color.test}}.
}
