% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.acf.R
\name{show.acf}
\alias{show.acf}
\title{Scatter Plots Illustrating Autoorrelation}
\usage{
show.acf(in.object, prow = 3, pcol = 5)
}
\arguments{
\item{in.object}{Time series or time series data objects (created by function code{\link{tsd}}).}

\item{prow}{Number of rows in the matrix of scatter plots to be displayed (default is 3). See mfrow in \code{\link{par}}.}

\item{pcol}{Number of columns in the matrix of scatter plots to be displayed (default is 5). See mfcol in \code{\link{par}}.}
}
\description{
Displays a sequence of scatter plots for an input time series, illustrating the autocorrelation between observations seperated by time lags between  (by default) 0 and 13, and the summary ACF function.
}
\examples{
plot(spot.tsd)
show.acf(spot.tsd)

}
\seealso{
\code{\link{par}}
}
