% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.true.acfpacf.R
\name{show.true.acfpacf}
\alias{show.true.acfpacf}
\title{True ACF and PACF Functions}
\usage{
show.true.acfpacf(model, nacf = 24, print.table = FALSE)
}
\arguments{
\item{model}{Model Specified ARMA model.}

\item{nacf}{Number of ACF/APCF lags to be computed and plotted (default is 24).}

\item{print.table}{If true, a ttable of the functions is printed (default is FALSE).}
}
\description{
Given a specified ARMA model (and model parameters), this function will compute and plot the true ACF and PACF functions.
}
\examples{
show.true.acfpacf(model=list(ar=0.9))
show.true.acfpacf(model=list(ar=-0.9))
show.true.acfpacf(model=list(ar=c(0.9, -0.6)))
show.true.acfpacf(model=list(ar=c(-1.5, -0.6)))
show.true.acfpacf(model=list(ar=c(1.5, -0.6)))
show.true.acfpacf(model=list(ma=0.9))
show.true.acfpacf(model=list(ma=c(0.9, 0.7)))
show.true.acfpacf(model=list(ar=0.9, ma=.09))
show.true.acfpacf(model=list(ar=0.9, ma=-0.9))
show.true.acfpacf(model=list(ar=c(0.9, -0.5), ma=c(0.1, 0.1)))

}
