% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simnsa.tsd}
\alias{simnsa.tsd}
\title{Simulated Time Series A}
\format{A time series data object (created with function tsd).

 @source Simulated.}
\usage{
simnsa.tsd
}
\description{
Simulated data from a non-stationary ARIMA model. Monthly data from January 1985 to January 2001.
}
\keyword{datasets}
