% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simnsb.tsd}
\alias{simnsb.tsd}
\title{Simulated Time Series B}
\format{
A time series data object (created with function tsd).

 @source Simulated.
}
\usage{
simnsb.tsd
}
\description{
Simulated data from a non-stationary ARIMA model. Monthly data from January 1989 to July 2001.
}
\keyword{datasets}
