% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simnsc.tsd}
\alias{simnsc.tsd}
\title{Simulated Time Series C}
\format{A time series data object (created with function tsd).

 @source Simulated.}
\usage{
simnsc.tsd
}
\description{
Simulated data from a non-stationary ARIMA model. Monthly data from January 1991 to January 2007.
}
\keyword{datasets}
