% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simnsd.tsd}
\alias{simnsd.tsd}
\title{Simulated Time Series D}
\format{A time series data object (created with function tsd).

 @source Simulated.}
\usage{
simnsd.tsd
}
\description{
Simulated data from a non-stationary ARIMA model. Monthly data from January 1994 to December 2018.
}
\keyword{datasets}
