% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simnse.tsd}
\alias{simnse.tsd}
\title{Simulated Time Series E}
\format{A time series data object (created with function tsd).

 @source Simulated.}
\usage{
simnse.tsd
}
\description{
Simulated data from a non-stationary ARIMA model. Monthly data from January 1985 to December 2003.
}
\keyword{datasets}
