% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simnsf.tsd}
\alias{simnsf.tsd}
\title{Simulated Time Series F}
\format{
A time series data object (created with function tsd).

 @source Simulated.
}
\usage{
simnsf.tsd
}
\description{
Simulated data from a non-stationary ARIMA model. Monthly data from January 1991 to December 2015.
}
\keyword{datasets}
