% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simsta04.tsd}
\alias{simsta04.tsd}
\title{Simulated Time Series 04}
\format{A time series data object (created with function tsd).

 @source Simulated.}
\usage{
simsta04.tsd
}
\description{
Simulated data from a stationary ARMA model. Monthly data from January 1975 to August 1991.
}
\keyword{datasets}
