% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simsta05.tsd}
\alias{simsta05.tsd}
\title{Simulated Time Series 05}
\format{A time series data object (created with function tsd).

 @source Simulated.}
\usage{
simsta05.tsd
}
\description{
Simulated data from a stationary ARMA model. Monthly data from January 1995 to April 2003.
}
\keyword{datasets}
