% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simsta06.tsd}
\alias{simsta06.tsd}
\title{Simulated Time Series 06}
\format{
A time series data object (created with function tsd).

 @source Simulated.
}
\usage{
simsta06.tsd
}
\description{
Simulated data from a stationary ARMA model. Monthly data from January 1996 to August 2012.
}
\keyword{datasets}
