% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simsta09.tsd}
\alias{simsta09.tsd}
\title{Simulated Time Series 09}
\format{A time series data object (created with function tsd).

 @source Simulated.}
\usage{
simsta09.tsd
}
\description{
Simulated data from a stationary ARMA model. Monthly data from January 1986 to June 1998.
}
\keyword{datasets}
