% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simsta11.tsd}
\alias{simsta11.tsd}
\title{Simulated Time Series 11}
\format{A time series data object (created with function tsd).

 @source Simulated.}
\usage{
simsta11.tsd
}
\description{
Simulated data from a stationary ARMA model. Monthly data from January 1988 to August 2004.
}
\keyword{datasets}
