% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simsta13.tsd}
\alias{simsta13.tsd}
\title{Simulated Time Series 13}
\format{A time series data object (created with function tsd).

 @source Simulated.}
\usage{
simsta13.tsd
}
\description{
Simulated data from a stationary ARMA model. Monthly data from January 1990 to August 2006.
}
\keyword{datasets}
