% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simsta15.tsd}
\alias{simsta15.tsd}
\title{Simulated Time Series 15}
\format{
A time series data object (created with function tsd).

 @source Simulated.
}
\usage{
simsta15.tsd
}
\description{
Simulated data from a stationary ARMA model. Monthly data from January 1992 to August 2008.
}
\keyword{datasets}
