% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spot.tsd}
\alias{spot.tsd}
\title{Wolfer Sunspot Numbers}
\format{A time series data object (created with function tsd).}
\source{
Used as an example in Chapter 10 Box and Jenkins (1970).
}
\usage{
spot.tsd
}
\description{
Wolfer sunspot numbers between 1770 and 1869 (100 observations)
}
\keyword{datasets}
