% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.cumsum.R
\name{ts.cumsum}
\alias{ts.cumsum}
\title{Time Series Cumulative Sum}
\usage{
ts.cumsum(w, lag = 1, start)
}
\arguments{
\item{w}{is the input time series.}

\item{lag}{allows integration at seasonal lags (lag=1 is the default).}

\item{start}{allows specification of starting value for the integration. If not specifies, w[1] is used.}
}
\description{
Computes the cumulative sum (integration) of a time series with optional lagging. The function is the inverse of differencing (possibly seasonal).
}
\examples{
ozoneRegdat <- read.csv(RTseriesExtDataPath('ozoneRegdat.csv'))
ozo.xreg <- cbind(ozoneRegdat[,2], ts.cumsum(ozoneRegdat[, 3], lag=12),
                ts.cumsum(ozoneRegdat[, 4], lag=12))
esti(ozone.tsd, model=model.pdq(D=1, q=1, Q=1, period=12),
                y.range=c(0.1, 9), xreg.in=ozo.xreg, gamma=0)

}
