% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsd.R
\name{tsd}
\alias{tsd}
\title{Create a Time Series Data Object}
\usage{
tsd(
  data.ts,
  data.title = deparse(substitute(data.ts)),
  time.units = "Time",
  response.units = deparse(substitute(data.ts))
)
}
\arguments{
\item{data.ts}{A time series object (created with the R \code{\link{ts}} function). This data should contains start date, end date and frequency.}

\item{data.title}{Optional. Title for the data. Default is the name of the data.ts object.}

\item{time.units}{Optional. Time units of the data. Default is 'Time'.}

\item{response.units}{Optional. The units of time for the time series data. Default is the name of the data.ts object.}
}
\value{
This function returns a time series data object that can be used as input to the \code{\link{iden}}, \code{\link{esti}}, and other RTsereis functions.
}
\description{
Creates a time-series data object that contains a time series and related information, e.g. data structure, title and the units. Provide input for esti and iden.
}
\examples{
savings.rate.ts <- ts(savings.rate, frequency=4, start=1955)
savings.rate.tsd <- tsd(savings.rate.ts, data.title='Change in Inventory',
     response.units='Billions of Dollars', time.units='Year')
iden(savings.rate.tsd)
}
\seealso{
\code{\link{iden}}, \code{\link{esti}}
}
